requirements.txt
requests
beautifulsoup4
twilio
python-dotenv